requirements.txt
amqp==2.5.0
billiard==3.6.0.0
boto3==1.9.206
botocore==1.12.206
celery==4.3.0
Django==2.2.4
django-celery-results==1.1.2
django-rest-framework==0.1.0
django-storages==1.7.1
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
docutils==0.14
jmespath==0.9.4
kombu==4.6.3
numpy==1.17.0
pandas==0.25.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
redis==3.3.6
rest-condition==1.0.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
